backend/requirements.txt
opencv-python
mediapipe
requests
python-dotenv
websockets